adds test for missing relay field
moves codec tests to their own module
